codeflix/admin-catalogo/infrastructure/src/main/java/com/glauco/codeflix/infrastructure/category/CategoryMySQLGateway.java
package com.glauco.codeflix.infrastructure.category;

import com.glauco.codeflix.domain.category.Category;
import com.glauco.codeflix.domain.category.CategoryGateway;
import com.glauco.codeflix.domain.category.CategoryID;
import com.glauco.codeflix.domain.category.CategorySeachQuery;
import com.glauco.codeflix.domain.pagination.Pagination;
import com.glauco.codeflix.infrastructure.category.persistence.CategoryJpaEntity;
import com.glauco.codeflix.infrastructure.category.persistence.CategoryRepository;
import org.springframework.stereotype.Service;

import java.util.Objects;
import java.util.Optional;

@Service
public class CategoryMySQLGateway implements CategoryGateway {

    private final CategoryRepository repository;

    public CategoryMySQLGateway(final CategoryRepository repository) {
        this.repository = Objects.requireNonNull(repository);
    }

    @Override
    public Category create(final Category aCategory) {
        return this.repository.save(CategoryJpaEntity.from(aCategory)).toAggregate();
    }

    @Override
    public void deleteById(CategoryID anId) {

    }

    @Override
    public Optional<Category> findById(CategoryID anId) {
        return Optional.empty();
    }

    @Override
    public Category update(Category aCategory) {
        return null;
    }

    @Override
    public Pagination<Category> findAll(CategorySeachQuery aQuery) {
        return null;
    }
}
